# store/copr: set concurrency to partition number for limit statements