--- a/distsql/request_builder.go
+++ b/distsql/request_builder.go
@@ -159,7 +159,11 @@ func (builder *RequestBuilder) SetDAGRequest(dag *tipb.DAGRequest) *RequestBuild
 		// When the DAG is just simple scan and small limit, set concurrency to 1 would be sufficient.
 		if execCnt == 2 {
 			if limit.Limit < estimatedRegionRowCount {
-				builder.Request.Concurrency = 1
+				if kr := builder.Request.KeyRanges; kr != nil {
+					builder.Request.Concurrency = kr.PartitionNum()
+				} else {
+					builder.Request.Concurrency = 1
+				}
 			}
 		}
 	}
@@ -273,9 +277,13 @@ func (*RequestBuilder) getKVPriority(sv *variable.SessionVars) int {
 // "Concurrency", "IsolationLevel", "NotFillCache", "TaskID", "Priority", "ReplicaRead",
 // "ResourceGroupTagger", "ResourceGroupName"
 func (builder *RequestBuilder) SetFromSessionVars(sv *variable.SessionVars) *RequestBuilder {
+	distsqlConcurrency := sv.DistSQLScanConcurrency()
 	if builder.Request.Concurrency == 0 {
-		// Concurrency may be set to 1 by SetDAGRequest
-		builder.Request.Concurrency = sv.DistSQLScanConcurrency()
+		// Concurrency unset.
+		builder.Request.Concurrency = distsqlConcurrency
+	} else if builder.Request.Concurrency > distsqlConcurrency {
+		// Concurrency is set in SetDAGRequest, check the upper limit.
+		builder.Request.Concurrency = distsqlConcurrency
 	}
 	replicaReadType := sv.GetReplicaRead()
 	if sv.StmtCtx.WeakConsistency {